modify Scala tests
